+++ supertokens_python/recipe/emailpassword/utils.py
@@ -17,6 +17,7 @@ from os import environ
 from re import fullmatch
 from typing import TYPE_CHECKING, Any, Awaitable, Callable, List, Union
 
+from ..emailverification.types import User as EmailVerificationUser
 from .interfaces import APIInterface, RecipeInterface
 from .types import InputFormField, NormalisedFormField, User
 
@@ -24,15 +25,16 @@ if TYPE_CHECKING:
     from .recipe import EmailPasswordRecipe
     from supertokens_python.supertokens import AppInfo
 
+from typing import Dict
+
 from httpx import AsyncClient
-from supertokens_python.recipe.emailverification.utils import \
-    InputEmailVerificationConfig
 from supertokens_python.recipe.emailverification.utils import \
     OverrideConfig as EmailVerificationOverrideConfig
 from supertokens_python.recipe.emailverification.utils import \
     ParentRecipeEmailVerificationConfig
 from supertokens_python.utils import get_filtered_list
 
+from ..emailverification.utils import ParentRecipeEmailVerificationConfig
 from .constants import (FORM_FIELD_EMAIL_ID, FORM_FIELD_PASSWORD_ID,
                         RESET_PASSWORD)
 
@@ -73,8 +75,8 @@ async def default_email_validator(value: str) -> Union[str, None]:
 
 
 def default_get_reset_password_url(
-        app_info: AppInfo) -> Callable[[User, Any], Awaitable[str]]:
-    async def func(_: User, __):
+        app_info: AppInfo) -> Callable[[User, Dict[str, Any]], Awaitable[str]]:
+    async def func(_: User, __: Dict[str, Any]):
         return app_info.website_domain.get_as_string_dangerous(
         ) + app_info.website_base_path.get_as_string_dangerous() + RESET_PASSWORD
 
@@ -82,8 +84,8 @@ def default_get_reset_password_url(
 
 
 def default_create_and_send_custom_email(
-        app_info: AppInfo) -> Callable[[User, str, Any], Awaitable]:
-    async def func(user: User, password_reset_url_with_token: str, _):
+        app_info: AppInfo) -> Callable[[User, str, Dict[str, Any]], Awaitable[None]]:
+    async def func(user: User, password_reset_url_with_token: str, _: Dict[str, Any]):
         if ('SUPERTOKENS_ENV' in environ) and (
                 environ['SUPERTOKENS_ENV'] == 'testing'):
             return
@@ -94,8 +96,7 @@ def default_create_and_send_custom_email(
                 'passwordResetURL': password_reset_url_with_token
             }
             async with AsyncClient() as client:
-                await client.post('https://api.supertokens.io/0/st/auth/password/reset', json=data,
-                                  headers={'api-version': '0'})
+                await client.post('https://api.supertokens.io/0/st/auth/password/reset', json=data, headers={'api-version': '0'}) # type: ignore
         except Exception:
             pass
 
@@ -151,14 +152,6 @@ def normalise_sign_up_form_fields(
     return normalised_form_fields
 
 
-def validate_and_normalise_sign_up_config(config=None) -> SignUpFeature:
-    if config is None:
-        config = {}
-    form_fields = normalise_sign_up_form_fields(
-        config['form_fields'] if 'form_fields' in config else None)
-    return SignUpFeature(form_fields)
-
-
 class SignInFeature:
     def __init__(self, form_fields: List[NormalisedFormField]):
         self.form_fields = form_fields
@@ -181,8 +174,8 @@ def validate_and_normalise_sign_in_config(
 class InputResetPasswordUsingTokenFeature:
     def __init__(self,
                  get_reset_password_url: Union[Callable[[
-                     User, Any], Awaitable[str]], None] = None,
-                 create_and_send_custom_email: Union[Callable[[User, str, Any], Awaitable], None] = None):
+                     User, Dict[str, Any]], Awaitable[str]], None] = None,
+                 create_and_send_custom_email: Union[Callable[[User, str, Dict[str, Any]], Awaitable[None]], None] = None):
         self.get_reset_password_url = get_reset_password_url
         self.create_and_send_custom_email = create_and_send_custom_email
 
@@ -191,13 +184,23 @@ class ResetPasswordUsingTokenFeature:
     def __init__(self,
                  form_fields_for_password_reset_form: List[NormalisedFormField],
                  form_fields_for_generate_token_form: List[NormalisedFormField],
-                 get_reset_password_url: Callable[[User, Any], Awaitable[str]],
-                 create_and_send_custom_email: Callable[[User, str, Any], Awaitable]):
+                 get_reset_password_url: Callable[[User, Dict[str, Any]], Awaitable[str]],
+                 create_and_send_custom_email: Callable[[User, str, Dict[str, Any]], Awaitable[None]]):
         self.form_fields_for_password_reset_form = form_fields_for_password_reset_form
         self.form_fields_for_generate_token_form = form_fields_for_generate_token_form
         self.get_reset_password_url = get_reset_password_url
         self.create_and_send_custom_email = create_and_send_custom_email
 
+
+class InputEmailVerificationConfig:
+    def __init__(self,
+                 get_email_verification_url: Union[Callable[[
+                     User, Any], Awaitable[str]], None] = None,
+                 create_and_send_custom_email: Union[Callable[[
+                     User, str, Any], Awaitable[None]], None] = None
+                 ):
+        self.get_email_verification_url = get_email_verification_url
+        self.create_and_send_custom_email = create_and_send_custom_email
 
 def validate_and_normalise_reset_password_using_token_config(app_info: AppInfo, sign_up_config: InputSignUpFeature,
                                                              config: InputResetPasswordUsingTokenFeature) -> ResetPasswordUsingTokenFeature:
@@ -214,10 +217,11 @@ def validate_and_normalise_reset_password_using_token_config(app_info: AppInfo, 
     return ResetPasswordUsingTokenFeature(form_fields_for_password_reset_form, form_fields_for_generate_token_form,
                                           get_reset_password_url, create_and_send_custom_email)
 
-
 def email_verification_create_and_send_custom_email(
-        recipe: EmailPasswordRecipe, create_and_send_custom_email):
-    async def func(user, link, user_context):
+        recipe: EmailPasswordRecipe, create_and_send_custom_email: Callable[[
+                     User, str, Dict[str, Any]], Awaitable[None]]) -> Callable[[
+                     EmailVerificationUser, str, Dict[str, Any]], Awaitable[None]]:
+    async def func(user: EmailVerificationUser, link: str, user_context: Dict[str, Any]):
         user_info = await recipe.recipe_implementation.get_user_by_id(user.user_id, user_context)
         if user_info is None:
             raise Exception('Unknown User ID provided')
@@ -227,9 +231,11 @@ def email_verification_create_and_send_custom_email(
 
 
 def email_verification_get_email_verification_url(
-        recipe: EmailPasswordRecipe, get_email_verification_url):
-    async def func(user, user_context):
-        user_info = await recipe.recipe_implementation.get_user_by_id(user.id, user_context)
+        recipe: EmailPasswordRecipe, get_email_verification_url: Callable[[
+                     User, Any], Awaitable[str]]) -> Callable[[
+                     EmailVerificationUser, Any], Awaitable[str]]:
+    async def func(user: EmailVerificationUser, user_context: Dict[str, Any]):
+        user_info = await recipe.recipe_implementation.get_user_by_id(user.user_id, user_context)
         if user_info is None:
             raise Exception('Unknown User ID provided')
         return await get_email_verification_url(user_info, user_context)
@@ -238,17 +244,17 @@ def email_verification_get_email_verification_url(
 
 
 def validate_and_normalise_email_verification_config(
-        recipe: EmailPasswordRecipe, config: Union[InputEmailVerificationConfig, None], override: InputOverrideConfig):
+        recipe: EmailPasswordRecipe, config: Union[InputEmailVerificationConfig, None], override: InputOverrideConfig) -> ParentRecipeEmailVerificationConfig:
     create_and_send_custom_email = None
     get_email_verification_url = None
     if config is None:
         config = InputEmailVerificationConfig()
     if config.create_and_send_custom_email is not None:
         create_and_send_custom_email = email_verification_create_and_send_custom_email(recipe,
-                                                                                       config.create_and_send_custom_email)
+                                        config.create_and_send_custom_email)
     if config.get_email_verification_url is not None:
         get_email_verification_url = email_verification_get_email_verification_url(recipe,
-                                                                                   config.get_email_verification_url)
+                                        config.get_email_verification_url)
     return ParentRecipeEmailVerificationConfig(
         get_email_for_user_id=recipe.get_email_for_user_id,
         create_and_send_custom_email=create_and_send_custom_email,
@@ -299,11 +305,7 @@ def validate_and_normalise_user_input(recipe: EmailPasswordRecipe, app_info: App
         override = InputOverrideConfig()
     if reset_password_using_token_feature is None:
         reset_password_using_token_feature = InputResetPasswordUsingTokenFeature()
-    email_verification_feature = validate_and_normalise_email_verification_config(
-        recipe,
-        email_verification_feature,
-        override
-    )
+
     if sign_up_feature is None:
         sign_up_feature = InputSignUpFeature()
     return EmailPasswordConfig(
@@ -316,6 +318,10 @@ def validate_and_normalise_user_input(recipe: EmailPasswordRecipe, app_info: App
             sign_up_feature,
             reset_password_using_token_feature
         ),
+        validate_and_normalise_email_verification_config(
+        recipe,
         email_verification_feature,
+        override
+    ),
         OverrideConfig(functions=override.functions, apis=override.apis)
     )