+++ supertokens_python/recipe/thirdparty/asyncio/__init__.py
@@ -11,7 +11,6 @@
 # WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations
 # under the License.
-from deprecated.classic import deprecated
 
 from supertokens_python.recipe.thirdparty.recipe import ThirdPartyRecipe
 
@@ -59,21 +58,6 @@ async def revoke_email_verification_tokens(user_id: str, user_context=None):
     email = await ThirdPartyRecipe.get_instance().get_email_for_user_id(user_id, user_context)
     return await ThirdPartyRecipe.get_instance().email_verification_recipe.recipe_implementation.revoke_email_verification_tokens(
         user_id, email, user_context)
-
-
-@deprecated(reason="Use supertokens_python.get_user_oldest_first(...) function instead IF using core version >= 3.5")
-async def get_users_oldest_first(limit: int = None, next_pagination: str = None):
-    return await ThirdPartyRecipe.get_instance().recipe_implementation.get_users_oldest_first(limit, next_pagination)
-
-
-@deprecated(reason="Use supertokens_python.get_users_newest_first(...) function instead IF using core version >= 3.5")
-async def get_users_newest_first(limit: int = None, next_pagination: str = None):
-    return await ThirdPartyRecipe.get_instance().recipe_implementation.get_users_newest_first(limit, next_pagination)
-
-
-@deprecated(reason="Use supertokens_python.get_user_count(...) function instead IF using core version >= 3.5")
-async def get_user_count():
-    return await ThirdPartyRecipe.get_instance().recipe_implementation.get_user_count()
 
 
 async def get_user_by_id(user_id: str, user_context=None):